+++ FireAxe.GUI/ViewModels/AddonNodeViewModelDesign.cs
@@ -2,10 +2,19 @@
 
 namespace FireAxe.ViewModels;
 
-public class AddonNodeViewModelDesign : AddonNodeViewModel
+internal class AddonNodeViewModelDesign : AddonNodeViewModel
 {
     public AddonNodeViewModelDesign() : base(DesignHelper.CreateTestAddonNode())
     {
-
+        var addon = Addon!;
+        var addonRoot = addon.Root;
+        
+        // dependencies
+        for (int i = 0; i < 3; i++)
+        {
+            var dependency = AddonNode.Create<AddonNode>(addonRoot);
+            dependency.Name = $"test_{DesignHelper.GenerateRandomString(5, 15)}";
+            addon.AddDependentAddon(dependency.Id);
+        }
     }
 }